+++ src/orquestra/sdk/_base/_logs/_interfaces.py
@@ -6,6 +6,7 @@ Logs-related interfaces.
 """
 import typing as t
 from dataclasses import dataclass
+from enum import Enum
 
 from orquestra.sdk.schema.ir import TaskInvocationId
 from orquestra.sdk.schema.workflow_run import WorkflowRunId
@@ -39,6 +40,35 @@ class WorkflowLogs:
     ``pip install --uprade orquestra-sdk`` or report your use case to the SDK Team at
     Zapata Computing.
     """
+
+    def get_log_type(
+        self, log_type
+    ) -> t.Union[t.Mapping[TaskInvocationId, t.Sequence[str]], t.Sequence[str]]:
+        """
+        Return the specified log type.
+
+        This method wraps the regular attribute getters in order to allow parametrised
+        access to individual log types.
+        """
+        if log_type == self.WorkflowLogTypeName.PER_TASK:
+            return self.per_task
+        elif log_type == self.WorkflowLogTypeName.SYSTEM:
+            return self.system
+        elif log_type == self.WorkflowLogTypeName.ENV_SETUP:
+            return self.env_setup
+        elif log_type == self.WorkflowLogTypeName.OTHER:
+            return self.other
+        raise ValueError(f"Unknown workflow log type '{log_type}'.")
+
+    class WorkflowLogTypeName(Enum):
+        """
+        Enum for specifying the individual types of Workflow log.
+        """
+
+        PER_TASK = "per_task"
+        SYSTEM = "system"
+        ENV_SETUP = "env_setup"
+        OTHER = "other"
 
 
 class LogReader(t.Protocol):