+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 
 🚨 *Breaking Changes*
 * Compute Engine is now the default when logging in
+* `orq wf logs` now supports `--task`, `--system` and `--env-setup` flags to control which logs are shown / downloaded. If none of these flags are set, the user is prompted for a choice of log types, with the default being `all`.
 * `InlineImport()` is now always default `source_import` for tasks
 
 🔥 *Features*